resolve DownloadsTableModel import warning
