+++ src/app/ngrx/index.ts
@@ -1,7 +1,7 @@
 export { AccountModel } from './reducers/account.reducer';
 export { PodcastModel } from './reducers/podcast.reducer';
 export { EpisodeModel, EPISODE_PAGE_SIZE } from './reducers/episode.reducer';
-export { RouterModel,
+export { RouterModel, DownloadsTableModel,
   INTERVAL_MONTHLY, INTERVAL_WEEKLY, INTERVAL_DAILY, INTERVAL_HOURLY, IntervalModel, IntervalList,
   CHARTTYPE_PODCAST, CHARTTYPE_EPISODES, CHARTTYPE_STACKED, ChartType,
   METRICSTYPE_DOWNLOADS, METRICSTYPE_DEMOGRAPHICS, METRICSTYPE_TRAFFICSOURCES, MetricsType, getMetricsProperty } from './reducers/models';
