+++ src/app/ngrx/reducers/models/downloads-table.model.ts
@@ -1,0 +1,12 @@
+import { TimeseriesDatumModel } from 'ngx-prx-styleguide';
+
+export interface DownloadsTableModel {
+  title: string;
+  publishedAt?: Date;
+  color: string;
+  id?: number;
+  downloads: TimeseriesDatumModel[];
+  totalForPeriod: number;
+  allTimeDownloads?: number;
+  charted: boolean;
+}
