+++ src/app/ngrx/reducers/models/index.ts
@@ -3,3 +3,4 @@ export { METRICSTYPE_DOWNLOADS, METRICSTYPE_DEMOGRAPHICS, METRICSTYPE_TRAFFICSOU
 export { INTERVAL_MONTHLY, INTERVAL_WEEKLY, INTERVAL_DAILY, INTERVAL_HOURLY, IntervalModel, IntervalList } from './interval.model';
 export { CHARTTYPE_PODCAST, CHARTTYPE_EPISODES, CHARTTYPE_STACKED, ChartType } from './chart.type';
 export { RouterModel } from './router.model';
+export { DownloadsTableModel } from './downloads-table.model';
